pyproject.toml
[tool.poetry]
name = "myrpl-cli"
version = "0.1.0"
description = "A script to fetch and save course activities from myrpl.ar"
authors = ["tcorzo"]

[tool.poetry.scripts]
myrpl = "myrpl_cli.main:main"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-dotenv = "^1.0.1"

# myrpl test dependencies:
timeout-decorator = "^0.5.0"
pytest = "^8.2.2"
tqdm = "^4.66.4"
requests-toolbelt = "^1.0.0"
keyring = "^25.2.1"
keyrings-cryptfile = "^1.3.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
